% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_proj_dir}
\alias{create_proj_dir}
\title{Function to create new project directory if it doesn't already exist.}
\usage{
create_proj_dir(directory)
}
\arguments{
\item{directory}{character vector designating the path to the new project directory}
}
\description{
Function to create new project directory if it doesn't already exist.
}
