% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draft_cc}
\alias{draft_cc}
\title{Function to draft files for project}
\usage{
draft_cc(file, template, metadata, package = NULL, create_dir, edit = FALSE)
}
\description{
Function to draft files for project
}
