% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report.R
\name{html_report_format}
\alias{html_report_format}
\title{Function that accepts arguments and formats for passing to rmarkdown/bookdown build functions.}
\usage{
html_report_format(
  template_name,
  template_path,
  template_dependencies,
  pandoc_args,
  ...
)
}
\arguments{
\item{template_name}{name of the template}

\item{template_path}{path to the html template, relative to the package 'inst' folder}

\item{template_dependencies}{list containing an html_dependency_* function call
that inserts desired .css and .js files}

\item{pandoc_args}{character vector containing pandoc output options}

\item{...}{additional arguments to be passed to the rmarkdown::html_document
or bookdown::html_document2 function}
}
\description{
Function that accepts arguments and formats for passing to rmarkdown/bookdown build functions.
}
