% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_functions.R
\name{new_user}
\alias{new_user}
\title{Function to create a new user file that will be used as defaults when creating a new project.}
\usage{
new_user(project_dir = NULL, name = NULL, email = NULL)
}
\arguments{
\item{project_dir}{Character vector that defines where}
}
\description{
Function to create a new user file that will be used as defaults when creating a new project.
}
