% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_id}
\alias{proj_id}
\title{Function to generate unique project id}
\usage{
proj_id(name, existing_ids)
}
\arguments{
\item{name}{character vector of which the first letter will be used in conjuction with a randomly sampled number to create a unique id.}

\item{existing_ids}{vector of existing project ids taken from}
}
\description{
Function to generate unique project id
}
