% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_functions.R
\name{project_create}
\alias{project_create}
\title{Function to create a new project.}
\usage{
project_create(
  project_folder = "~/",
  start_date = as.character(Sys.Date()),
  client_name = NULL,
  client_email = NULL,
  statistician_name = NULL,
  statistician_email = NULL,
  programmer = NULL,
  version = "1.0",
  department = NULL,
  project_title = NULL,
  output_format = "html_report",
  notes = ""
)
}
\arguments{
\item{project_folder}{Directory where project files are stored (usually a network drive location).}

\item{start_date}{Start date of project. Defaults to today's date.}

\item{client_name}{Name of PI/Client.}

\item{client_email}{Email of PI/Client.}

\item{statistician_name}{Name of lead statistician.}

\item{statistician_email}{Email of lead statistician.}

\item{version}{Optional version number. Defaults to version 1.0.}

\item{department}{PI/Client department affiliation.}

\item{project_title}{Optional title/description of project.}

\item{output_format}{Sets the output format for statistical reports. Options include
"html" (default) or "pdf." Output format can be
changed in YAML header of Rmarkdown document.}
}
\description{
Function to create a new project.
}
