% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yaml_maker}
\alias{yaml_maker}
\title{Function to generate YAML header in project files}
\usage{
yaml_maker(
  start_date = "`r format(Sys.time(), \\"\%Y/\%m/\%d\\")`",
  version = "1.0",
  dept = "GUKI",
  statistician_name = NULL,
  statistician_email = NULL,
  client_name = NULL,
  client_email = NULL,
  project_id = NULL,
  project_title = NULL,
  output_format = "html_report",
  ...
)
}
\description{
Function to generate YAML header in project files
}
